bot/requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
asyncio==3.4.3
pytest==7.4.3
pytest-asyncio==0.21.1